Update API Docs
 * Updated auto-generated API Docs (ApiGen) - refs #51
 * Added Google Analytics ID to API docs - refs #93
 * Added Google Custom Search Engine ID to API docs - refs #94

This should have been included in the v0.1 release or, failing that, I should have created a new v0.1.1 release to incorperate these changes to the API docs, but I'm not going to do all that just to add some stupid stuff to my API docs that aren't even related to the code. From now on any changes that are made after a release will be incorperated into a patch release.